Per code review, changing code structure to match Events view style
